separate watcher and watcherbitcoin stratumjob base now abstract class. Force serialize/unserialize pair for stratum job
