+++ src/bitcoin/Watcher.cc
@@ -23,9 +23,6 @@
  */
 #include "Watcher.h"
 
-#include "BitcoinUtils.h"
-#include "KafkaBitcoin.h"
-
 #include <arpa/inet.h>
 #include <cinttypes>
 
@@ -34,10 +31,6 @@
 #include <event2/bufferevent.h>
 #include <event2/listener.h>
 #include <glog/logging.h>
-
-#include <chainparams.h>
-#include <utilstrencodings.h>
-#include <hash.h>
 
 
 static
@@ -62,6 +55,7 @@ bool tryReadLine(string &line, struct bufferevent *bufev) {
 
 static
 bool resolve(const string &host, struct	in_addr *sin_addr) {
+  LOG(INFO) << "resolve " << host.c_str();
   struct evutil_addrinfo *ai = NULL;
   struct evutil_addrinfo hints_in;
   memset(&hints_in, 0, sizeof(evutil_addrinfo));
@@ -74,11 +68,11 @@ bool resolve(const string &host, struct	in_addr *sin_addr) {
   // TODO: use non-blocking to resolve hostname
   int err = evutil_getaddrinfo(host.c_str(), NULL, &hints_in, &ai);
   if (err != 0) {
-    LOG(ERROR) << "evutil_getaddrinfo err: " << err << ", " << evutil_gai_strerror(err);
+    LOG(ERROR) << "[" << host.c_str() << "] evutil_getaddrinfo err: " << err << ", " << evutil_gai_strerror(err);
     return false;
   }
   if (ai == NULL) {
-    LOG(ERROR) << "evutil_getaddrinfo res is null";
+    LOG(ERROR) << "[" << host.c_str() << "] evutil_getaddrinfo res is null";
     return false;
   }
 
@@ -99,40 +93,11 @@ bool resolve(const string &host, struct	in_addr *sin_addr) {
   return true;
 }
 
-static
-int32_t getBlockHeightFromCoinbase(const string &coinbase1) {
-  // https://github.com/bitcoin/bips/blob/master/bip-0034.mediawiki
-  const string a = coinbase1.substr(86, 2);
-  const string b = coinbase1.substr(88, 2);
-  const string c = coinbase1.substr(90, 2);
-  const string heightHex = c + b + a;  // little-endian
-
-  return (int32_t)strtol(heightHex.c_str(), nullptr, 16);
-}
-
-
-//
-// input  : 89c2f63dfb970e5638aa66ae3b7404a8a9914ad80328e9fe0000000000000000
-// output : 00000000000000000328e9fea9914ad83b7404a838aa66aefb970e5689c2f63d
-static
-string convertPrevHash(const string &prevHash) {
-  assert(prevHash.length() == 64);
-  string hash;
-  for (int i = 7; i >= 0; i--) {
-    uint32_t v = (uint32_t)strtoul(prevHash.substr(i*8, 8).c_str(), nullptr, 16);
-    hash.append(Strings::Format("%08x", v));
-  }
-  return hash;
-}
-
-
-
 ///////////////////////////////// ClientContainer //////////////////////////////
-ClientContainer::ClientContainer(const string &kafkaBrokers)
-:running_(true), kafkaBrokers_(kafkaBrokers),
-kafkaProducer_(kafkaBrokers_.c_str(), KAFKA_TOPIC_RAWGBT, 0/* partition */),
-kafkaStratumJobConsumer_(kafkaBrokers_.c_str(), KAFKA_TOPIC_STRATUM_JOB, 0/*patition*/),
-poolStratumJob_(nullptr)
+ClientContainer::ClientContainer(const string &kafkaBrokers, const char* consumerTopic, const char* producerTopic)
+  : running_(true), kafkaBrokers_(kafkaBrokers)
+  , kafkaProducer_(kafkaBrokers_.c_str(), producerTopic, 0/* partition */)
+  , kafkaStratumJobConsumer_(kafkaBrokers_.c_str(), consumerTopic, 0/*patition*/)
 {
   base_ = event_base_new();
   assert(base_ != nullptr);
@@ -140,14 +105,6 @@ poolStratumJob_(nullptr)
 
 ClientContainer::~ClientContainer() {
   event_base_free(base_);
-}
-
-boost::shared_lock<boost::shared_mutex> ClientContainer::getPoolStratumJobReadLock() {
-  return boost::shared_lock<boost::shared_mutex>(stratumJobMutex_);
-}
-
-const StratumJobBitcoin * ClientContainer::getPoolStratumJob() {
-  return poolStratumJob_;
 }
 
 void ClientContainer::runThreadStratumJobConsume() {
@@ -175,68 +132,31 @@ void ClientContainer::runThreadStratumJobConsume() {
 }
 
 void ClientContainer::consumeStratumJob(rd_kafka_message_t *rkmessage) {
-    // check error
-    if (rkmessage->err) {
-      if (rkmessage->err == RD_KAFKA_RESP_ERR__PARTITION_EOF) {
-        // Reached the end of the topic+partition queue on the broker.
-        // Not really an error.
-        //      LOG(INFO) << "consumer reached end of " << rd_kafka_topic_name(rkmessage->rkt)
-        //      << "[" << rkmessage->partition << "] "
-        //      << " message queue at offset " << rkmessage->offset;
-        // acturlly
-        return;
-      }
-    
-      LOG(ERROR) << "consume error for topic " << rd_kafka_topic_name(rkmessage->rkt)
-      << "[" << rkmessage->partition << "] offset " << rkmessage->offset
-      << ": " << rd_kafka_message_errstr(rkmessage);
-    
-      if (rkmessage->err == RD_KAFKA_RESP_ERR__UNKNOWN_PARTITION ||
-          rkmessage->err == RD_KAFKA_RESP_ERR__UNKNOWN_TOPIC) {
-        LOG(FATAL) << "consume fatal";
-      }
+  // check error
+  if (rkmessage->err) {
+    if (rkmessage->err == RD_KAFKA_RESP_ERR__PARTITION_EOF) {
+      // Reached the end of the topic+partition queue on the broker.
+      // Not really an error.
+      //      LOG(INFO) << "consumer reached end of " << rd_kafka_topic_name(rkmessage->rkt)
+      //      << "[" << rkmessage->partition << "] "
+      //      << " message queue at offset " << rkmessage->offset;
+      // acturlly
       return;
     }
   
-    StratumJobBitcoin *sjob = new StratumJobBitcoin();
-    bool res = sjob->unserializeFromJson((const char *)rkmessage->payload,
-                                         rkmessage->len);
-    if (res == false) {
-      LOG(ERROR) << "unserialize stratum job fail";
-      delete sjob;
-      return;
+    LOG(ERROR) << "consume error for topic " << rd_kafka_topic_name(rkmessage->rkt)
+    << "[" << rkmessage->partition << "] offset " << rkmessage->offset
+    << ": " << rd_kafka_message_errstr(rkmessage);
+  
+    if (rkmessage->err == RD_KAFKA_RESP_ERR__UNKNOWN_PARTITION ||
+        rkmessage->err == RD_KAFKA_RESP_ERR__UNKNOWN_TOPIC) {
+      LOG(FATAL) << "consume fatal";
     }
-    // make sure the job is not expired.
-    if (jobId2Time(sjob->jobId_) + 60 < time(nullptr)) {
-      LOG(ERROR) << "too large delay from kafka to receive topic 'StratumJob'";
-      delete sjob;
-      return;
-    }
+    return;
+  }
 
-    DLOG(INFO) << "[POOL] stratum job received, height: " << sjob->height_
-              << ", prevhash: " << sjob->prevHash_.ToString()
-              << ", nBits: " << sjob->nBits_;
-
-    {
-      // get a write lock before change this->poolStratumJob_
-      // it will unlock by itself in destructor.
-      boost::unique_lock<boost::shared_mutex> writeLock(stratumJobMutex_);
-
-      uint256 oldPrevHash;
-
-      if (poolStratumJob_ != nullptr) {
-        oldPrevHash = poolStratumJob_->prevHash_;
-        delete poolStratumJob_;
-      }
-
-      poolStratumJob_ = sjob;
-
-      if (oldPrevHash != sjob->prevHash_) {
-        LOG(INFO) << "[POOL] prev block changed, height: " << sjob->height_
-                  << ", prevhash: " << sjob->prevHash_.ToString()
-                  << ", nBits: " << sjob->nBits_;
-      }
-    }
+  string str((const char*)rkmessage->payload, rkmessage->len);
+  consumeStratumJobInternal(str);
 }
 
 void ClientContainer::run() {
@@ -301,8 +221,7 @@ bool ClientContainer::init() {
 // do add pools before init()
 bool ClientContainer::addPools(const string &poolName, const string &poolHost,
                                const int16_t poolPort, const string &workerName) {
-  auto ptr = new PoolWatchClient(base_, this,
-                                 poolName, poolHost, poolPort, workerName);
+  auto ptr = createPoolWatchClient(base_, poolName, poolHost, poolPort, workerName);
   if (!ptr->connect()) {
     return false;
   }
@@ -311,51 +230,10 @@ bool ClientContainer::addPools(const string &poolName, const string &poolHost,
   return true;
 }
 
-bool ClientContainer::makeEmptyGBT(int32_t blockHeight, uint32_t nBits,
-                                   const string &blockPrevHash,
-                                   uint32_t blockTime, uint32_t blockVersion) {
-
-  // generate empty GBT
-  string gbt;
-  gbt += Strings::Format("{\"result\":{");
-
-  gbt += Strings::Format("\"previousblockhash\":\"%s\"", blockPrevHash.c_str());
-  gbt += Strings::Format(",\"height\":%d", blockHeight);
-  const CChainParams& chainparams = Params();
-  gbt += Strings::Format(",\"coinbasevalue\":%" PRId64"",
-                         GetBlockReward(blockHeight, chainparams.GetConsensus()));
-  gbt += Strings::Format(",\"bits\":\"%08x\"", nBits);
-  const uint32_t minTime = blockTime - 60*10;  // just set 10 mins ago
-  gbt += Strings::Format(",\"mintime\":%" PRIu32"", minTime);
-  gbt += Strings::Format(",\"curtime\":%" PRIu32"", blockTime);
-  gbt += Strings::Format(",\"version\":%" PRIu32"", blockVersion);
-  gbt += Strings::Format(",\"transactions\":[]");  // empty transactions
-
-  gbt += Strings::Format("}}");
-
-  const uint256 gbtHash = Hash(gbt.begin(), gbt.end());
-
-  string sjob = Strings::Format("{\"created_at_ts\":%u,"
-                                "\"block_template_base64\":\"%s\","
-                                "\"gbthash\":\"%s\"}",
-                                (uint32_t)time(nullptr),
-                                EncodeBase64(gbt).c_str(),
-                                gbtHash.ToString().c_str());
-
-  // submit to Kafka
-  kafkaProducer_.produce(sjob.c_str(), sjob.length());
-
-  LOG(INFO) << "sumbit to Kafka, msg len: " << sjob.length();
-  LOG(INFO) << "empty gbt: " << gbt;
-
-  return true;
-}
-
 void ClientContainer::removeAndCreateClient(PoolWatchClient *client) {
   for (size_t i = 0; i < clients_.size(); i++) {
     if (clients_[i] == client) {
-      auto ptr = new PoolWatchClient(base_, this,
-                                     client->poolName_, client->poolHost_,
+      auto ptr = createPoolWatchClient(base_, client->poolName_, client->poolHost_,
                                      client->poolPort_, client->workerName_);
       ptr->connect();
       LOG(INFO) << "reconnect " << ptr->poolName_;
@@ -384,8 +262,7 @@ void ClientContainer::eventCallback(struct bufferevent *bev,
     client->state_ = PoolWatchClient::State::CONNECTED;
 
     // do subscribe
-    string s = Strings::Format("{\"id\":1,\"method\":\"mining.subscribe\""
-                               ",\"params\":[\"%s\"]}\n", BTCCOM_WATCHER_AGENT);
+    string s = container->createOnConnectedReplyString();
     client->sendData(s);
     return;
   }
@@ -414,8 +291,11 @@ PoolWatchClient::PoolWatchClient(struct event_base *base, ClientContainer *conta
                                  const string &poolName,
                                  const string &poolHost, const int16_t poolPort,
                                  const string &workerName)
-: container_(container), poolName_(poolName), poolHost_(poolHost),
-poolPort_(poolPort), workerName_(workerName)
+  : container_(container)
+  , poolName_(poolName)
+  , poolHost_(poolHost)
+  , poolPort_(poolPort)
+  , workerName_(workerName)
 {
   bev_ = bufferevent_socket_new(base, -1, BEV_OPT_CLOSE_ON_FREE);
   assert(bev_ != nullptr);
@@ -425,8 +305,6 @@ poolPort_(poolPort), workerName_(workerName)
                     ClientContainer::eventCallback, this);
   bufferevent_enable(bev_, EV_READ|EV_WRITE);
 
-  extraNonce1_ = 0;
-  extraNonce2Size_ = 0;
 
   state_ = INIT;
 
@@ -447,6 +325,8 @@ bool PoolWatchClient::connect() {
   if (!resolve(poolHost_, &sin.sin_addr)) {
     return false;
   }
+
+  LOG(INFO) << "Connection request to " << poolHost_ << ":" << poolPort_;
 
   // bufferevent_socket_connect(): This function returns 0 if the connect
   // was successfully launched, and -1 if an error occurred.
@@ -480,156 +360,3 @@ bool PoolWatchClient::handleMessage() {
   return false;
 }
 
-void PoolWatchClient::handleStratumMessage(const string &line) {
-  DLOG(INFO) << "<" << poolName_ << "> UpPoolWatchClient recv(" << line.size() << "): " << line;
-
-  JsonNode jnode;
-  if (!JsonNode::parse(line.data(), line.data() + line.size(), jnode)) {
-    LOG(ERROR) << "decode line fail, not a json string";
-    return;
-  }
-  JsonNode jresult = jnode["result"];
-  JsonNode jerror  = jnode["error"];
-  JsonNode jmethod = jnode["method"];
-
-  if (jmethod.type() == Utilities::JS::type::Str) {
-    JsonNode jparams  = jnode["params"];
-    auto jparamsArr = jparams.array();
-
-    if (jmethod.str() == "mining.notify") {
-      const string prevHash = convertPrevHash(jparamsArr[1].str());
-
-      if (lastPrevBlockHash_.empty()) {
-        lastPrevBlockHash_ = prevHash;  // first set prev block hash
-      }
-
-      // stratum job prev block hash changed
-      if (lastPrevBlockHash_ != prevHash) {
-
-        // block height in coinbase (BIP34)
-        const int32_t  blockHeight = getBlockHeightFromCoinbase(jparamsArr[2].str());
-        
-        // nBits, the encoded form of network target
-        const uint32_t nBits       = jparamsArr[6].uint32_hex();
-
-        // only for display, it will be replaced by current system time
-        uint32_t blockTime   = jparamsArr[7].uint32_hex();
-
-        // only for display, it will be replaced by current poolStratumJob's nVersion
-        uint32_t nVersion    = jparamsArr[5].uint32_hex();
-
-        lastPrevBlockHash_ = prevHash;
-        LOG(INFO) << "<" << poolName_ << "> prev block changed, height: " << blockHeight
-                               << ", prev_hash: " << prevHash
-                               << ", block_time: " << blockTime
-                               << ", nBits: " << nBits
-                               << ", nVersion: " << nVersion;
-
-        //////////////////////////////////////////////////////////////////////////
-        // To ensure the external job is not deviation from the blockchain.
-        // 
-        // eg. a Bitcoin pool may receive a Bitcoin Cash job from a external
-        // stratum server, because the stratum server is automatic switched
-        // between Bitcoin and Bitcoin Cash depending on profit.
-        //////////////////////////////////////////////////////////////////////////
-        {
-          // get a read lock before lookup this->poolStratumJob_
-          // it will unlock by itself in destructor.
-          auto readLock = container_->getPoolStratumJobReadLock();
-          const StratumJobBitcoin *poolStratumJob = container_->getPoolStratumJob();
-
-          if (poolStratumJob == nullptr) {
-            LOG(WARNING) << "<" << poolName_ << "> discard the job: pool stratum job is empty";
-            return;
-          }
-
-          if (blockHeight == poolStratumJob->height_) {
-            LOG(INFO) << "<" << poolName_ << "> discard the job: height is same as pool."
-                                      << " pool height: " << poolStratumJob->height_
-                                      << ", the job height: " << blockHeight;
-            return;
-          }
-
-          if (blockHeight != poolStratumJob->height_ + 1) {
-            LOG(WARNING) << "<" << poolName_ << "> discard the job: height jumping too much."
-                                      << " pool height: " << poolStratumJob->height_
-                                      << ", the job height: " << blockHeight;
-            return;
-          }
-
-          if (nBits != poolStratumJob->nBits_) {
-            LOG(WARNING) << "<" << poolName_ << "> discard the job: nBits different from pool job."
-                                      << " pool nBits: " << poolStratumJob->nBits_
-                                      << ", the job nBits: " << nBits;
-            return;
-          }
-
-          // the block time from other pool may have a deviation with the current time.
-          // so replaced it by current system time.
-          blockTime = (uint32_t)time(nullptr);
-
-          // the nVersion from other pool may have some flags that we don't want.
-          // so replaced it by current poolStratumJob's.
-          nVersion = poolStratumJob->nVersion_;
-        }
-
-        container_->makeEmptyGBT(blockHeight, nBits, prevHash, blockTime, nVersion);
-
-      }
-    }
-    else {
-      // ignore other messages
-    }
-    return;
-  }
-
-  if (state_ == AUTHENTICATED) {
-    //
-    // {"error": null, "id": 2, "result": true}
-    //
-    if (jerror.type()  != Utilities::JS::type::Null ||
-        jresult.type() != Utilities::JS::type::Bool ||
-        jresult.boolean() != true) {
-      LOG(ERROR) << poolName_ <<  " auth fail";
-    }
-    return;
-  }
-
-  if (state_ == CONNECTED) {
-    //
-    // {"id":1,"result":[[["mining.set_difficulty","01000002"],
-    //                    ["mining.notify","01000002"]],"01000002",8],"error":null}
-    //
-    if (jerror.type() != Utilities::JS::type::Null) {
-      LOG(ERROR) << "<" << poolName_ << "> json result is null, err: " << jerror.str();
-      return;
-    }
-    std::vector<JsonNode> resArr = jresult.array();
-    if (resArr.size() < 3) {
-      LOG(ERROR) << "<" << poolName_ << "> result element's number is less than 3: " << line;
-      return;
-    }
-
-    extraNonce1_     = resArr[1].uint32_hex();
-    extraNonce2Size_ = resArr[2].uint32();
-    LOG(INFO) << "<" << poolName_ << "> extraNonce1: " << extraNonce1_
-    << ", extraNonce2 Size: " << extraNonce2Size_;
-
-    // subscribe successful
-    state_ = SUBSCRIBED;
-
-    // do mining.authorize
-    string s = Strings::Format("{\"id\": 1, \"method\": \"mining.authorize\","
-                               "\"params\": [\"%s\", \"\"]}\n",
-                               workerName_.c_str());
-    sendData(s);
-    return;
-  }
-
-  if (state_ == SUBSCRIBED && jresult.boolean() == true) {
-    // authorize successful
-    state_ = AUTHENTICATED;
-    LOG(INFO) << "<" << poolName_ << "> auth success, name: \"" << workerName_ << "\"";
-    return;
-  }
-}
