+++ src/sia/StratumSia.h
@@ -32,6 +32,7 @@ class StratumJobSia : public StratumJob
 public:
 
 public:
+  string serializeToJson() const;
   bool unserializeFromJson(const char *s, size_t len) override;
   uint32 jobTime() const override { return nTime_; }
 };
